Replace lzcomp with lzpcompress
