+++ tools/.gitignore
@@ -3,7 +3,7 @@ bpp2png
 bspcomp
 fine_print
 gfx
-lzcomp
+lzpcompress
 make_patch
 png_dimensions
 pokemon_animation
